ViewModel added and packages re worked
