+++ app/src/main/java/com/example/strength4mom/MainActivity.kt
@@ -1,40 +1,13 @@
 package com.example.strength4mom
-
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ExpandLess
-import androidx.compose.material.icons.filled.ExpandMore
-import androidx.compose.material3.Card
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.res.stringResource
-import com.example.strength4mom.data.Exo
-import com.example.strength4mom.data.exos
-import com.example.strength4mom.ui.theme.Strength4MomTheme
-import kotlinx.coroutines.flow.MutableStateFlow
+import com.example.strength4mom.ui.theme.theme.Strength4MomTheme
+import com.example.strength4mom.ui.theme.exo.StrengthApp
 
 
 class MainActivity : ComponentActivity() {
@@ -58,103 +31,3 @@ class MainActivity : ComponentActivity() {
 
 
 
-@Composable
-fun StrengthApp() {
-    LazyColumn() {
-        items(exos) {
-            ExoItem(
-                exo = it,
-                modifier = Modifier
-                    .padding(dimensionResource(R.dimen.padding_small))
-            )
-        }
-    }
-}
-
-
-
-/**
- * Composable displaying the list of exercises with the image and information
- *
- */
-@Composable
-fun ExoItem(
-    exo: Exo,
-    modifier: Modifier = Modifier
-) {
-
-//To added ViewModel once Module 4 is done.
-//    val expended = remember { MutableStateFlow(false) }
-//    val expendedValue = expended.collectAsState()
-
-    Card(modifier = modifier) {
-        Column(modifier = modifier) {
-            Row(modifier = modifier) {
-                ExoInfo(exoName = exo.name, exoSets = exo.numberSet, exoReps = exo.numberRep)
-            }
-        }
-    }
-}
-
-@Composable
-fun ExoInfo(
-    @StringRes exoName: Int,
-    exoSets: Int,
-    exoReps: Int,
-    modifier: Modifier = Modifier
-) {
-    var expended by remember { mutableStateOf(false) }
-
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-    ) {
-        Text(
-            text = stringResource(exoName),
-            style = MaterialTheme.typography.displayMedium
-        )
-        Row(
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = stringResource(R.string.sets, exoSets),
-                style = MaterialTheme.typography.displaySmall,
-                modifier = modifier
-                    .weight(1f)
-            )
-            Text(
-                text = stringResource(R.string.reps, exoReps),
-                style = MaterialTheme.typography.displaySmall,
-                modifier = modifier
-                    .weight(1f)
-            )
-            ExoButton(expended = expended, onClick = { expended = !expended })
-        }
-    }
-}
-
-
-//Added Gradle dependency to be able to use icons from https://fonts.google.com/icons
-//Added implementation("androidx.compose.material:material-icons-extended") to to :app build.gradle.kts file
-@Composable
-private fun ExoButton(
-    expended: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    IconButton(
-        onClick = onClick,
-        modifier = modifier
-    ) {
-        Icon(
-            imageVector = if (!expended) Icons.Filled.ExpandMore else Icons.Filled.ExpandLess,
-            contentDescription = stringResource(R.string.expand_button),
-            tint = MaterialTheme.colorScheme.secondary,
-        )
-    }
-}
-
-//ExoInfo
-//ExoButton
-//ExoImage
-//ExoDescription
