+++ app/src/main/java/com/example/strength4mom/ui/theme/exo/ExoUiState.kt
@@ -1,0 +1,8 @@
+package com.example.strength4mom.ui.theme.exo
+
+
+
+data class ExoUiState(
+    var title: String = "",
+    var expanded: Boolean = false,
+)
