+++ app/src/main/java/com/example/strength4mom/ui/theme/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.example.strength4mom.ui.theme
+package com.example.strength4mom.ui.theme.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
