+++ app/src/main/java/com/example/strength4mom/ui/theme/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.strength4mom.ui.theme
+package com.example.strength4mom.ui.theme.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
